Aligned parameter names between base or interface and implementors.  This sometimes meant modifying parameter name in base or interface to be more descriptive (replace "t" with "time" and "ts" with "times"). Removed unnecessary casts.

+++ src/Vts/Common/Extensions/DataExtensions.cs
@@ -53,7 +53,7 @@ namespace Vts.Extensions
                 double gaussDeviate = Math.Sqrt(-2 * Math.Log(uniformDeviate1)) *
                     Math.Cos(2 * Math.PI * uniformDeviate2);
                 
-                yield return (double) (d * (1 + noiseFraction * gaussDeviate));
+                yield return  (d * (1 + noiseFraction * gaussDeviate));
             }
 
         }
